crud funcionando

+++ src/main/java/br/com/strella/srv/empresa/adapter/output/repository/EmpresaAdapter.java
@@ -3,10 +3,13 @@ package br.com.strella.srv.empresa.adapter.output.repository;
 import br.com.strella.srv.empresa.adapter.output.repository.dto.EmpresaOutputDTO;
 import br.com.strella.srv.empresa.adapter.output.repository.dto.mapper.EmpresaMapperOutput;
 import br.com.strella.srv.empresa.domain.entity.Empresa;
+import br.com.strella.srv.empresa.domain.exception.ArquivoNotFoudException;
 import br.com.strella.srv.empresa.port.output.IEmpresaPort;
 import java.util.List;
+import java.util.logging.Logger;
 import org.hibernate.Hibernate;
 import org.hibernate.HibernateException;
+import org.springframework.boot.actuate.autoconfigure.health.HealthEndpointProperties.Logging;
 import org.springframework.data.domain.Pageable;
 
 public class EmpresaAdapter implements IEmpresaPort {
@@ -45,7 +48,12 @@ public class EmpresaAdapter implements IEmpresaPort {
 	public void deletarEmpresa(Empresa empresa) throws Exception {
 		try {
 			EmpresaOutputDTO empresaOutputDTO = EmpresaMapperOutput.INSTANCE.empresaToEmpresaOutputDTO(empresa);
-			empresaRepository.delete(empresaOutputDTO);
+			if(empresaRepository.existsById(empresa.getId()))
+				empresaRepository.delete(empresaOutputDTO);
+			else
+				throw new ArquivoNotFoudException("404", "empresa nao encontrada");
+
+
 		} catch (HibernateException exception) {
 			throw new Exception();
 		}
